# swagger.yaml
...
swagger: "2.0"
info:
  description: This is a API of SaratovNews
  version: 1.0.0
  title: Simple API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: virtserver.swaggerhub.com
basePath: /Ometh322/SararovNews/1.0.0
...
schemes:
- https
- http
paths:
  /newsFeed:
    get:
      tags:
      - newsFeed
      summary: List of All News, Events, Polls in News Feed
      operationId: listNewsFeed
      produces:
      - application/json
      - application/xml
      parameters:
      - name: numberOfObject
        in: query
        description: Number of Outputing Objects
        required: false
        type: integer
      - name: startIndex
        in: query
        description: Index of First Object, that need to be Outputed
        required: false
        type: integer
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/Record'
        "404":
          description: News Feed not found
  /news:
    get:
      tags:
      - news
      summary: List All News
      description: Returns the whole News
      operationId: listNews
      produces:
      - application/json
      - application/xml
      parameters: []
      responses:
        "200":
          description: Successful operation
          schema:
            type: array
            items:
              $ref: '#/definitions/News'
        "404":
          description: News not found
    put:
      tags:
      - news
      summary: Adds a new News post
      operationId: addNews
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - in: body
        name: body
        description: News object that needs to be added
        required: true
        schema:
          $ref: '#/definitions/News'
      responses:
        "201":
          description: Successful creating
          schema:
            $ref: '#/definitions/News'
        "405":
          description: Invalid Input
  /news/{id}:
    get:
      tags:
      - news
      summary: Find News by ID
      description: Returns a single News
      operationId: getNewsById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of News to return
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
          schema:
            $ref: '#/definitions/News'
        "400":
          description: Invalid ID supplied
        "404":
          description: News not found
    post:
      tags:
      - news
      summary: Update news by ID
      operationId: updateNews
      consumes:
      - application/json
      - application/xml
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: ID of News that need to be updated
        required: true
        type: integer
        format: int64
      - in: body
        name: body
        description: News object that needs to be updated
        required: true
        schema:
          $ref: '#/definitions/News'
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: News not found
        "405":
          description: Validation exception
    delete:
      tags:
      - news
      summary: Deletes a News post
      operationId: deleteNews
      produces:
      - application/json
      - application/xml
      parameters:
      - name: id
        in: path
        description: News ID to delete
        required: true
        type: integer
        format: int64
      responses:
        "200":
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: News not found
...
definitions:
  News:
    allOf:
    - $ref: '#/definitions/Record'
    - {}
...
  Comments:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      message:
        type: string
      publicationDate:
        type: string
        format: date-time
      numberOfLikes:
        type: integer
        format: int32
  Record:
    type: object
    properties:
      id:
        type: integer
        format: int64
      creatorId:
        type: integer
        format: int64
      title:
        type: string
      publicationDate:
        type: string
        format: date-time
      description:
        type: string
      numberOfComments:
        type: integer
        format: int32
      numberOfLikes:
        type: integer
        format: int32
      comments:
        type: array
        items:
          $ref: '#/definitions/Comments'
      typeOfRecord:
        type: string
        enum:
        - NEWS
        - EVENT
        - POLL
...